+++ app/src/main/java/cn/teachcourse/view/dialog/AppCompatActivityDialog.java
@@ -35,7 +35,7 @@ public class AppCompatActivityDialog extends AppCompatActivity {
     private View createDialog(TestDialogBean bean) {
         /**LinearLayout默认水平方向布局*/
         LinearLayout ll = new LinearLayout(this);
-        ll.setLayoutParams(new ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, DensityUtil.px2dip(this, 400)));
+        ll.setLayoutParams(new ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, (int)DensityUtil.px2dip(this, 400)));
         ll.setOrientation(LinearLayout.VERTICAL);
 
         TextView mTitle_tv = new TextView(this);
